+++ NEWS.md
@@ -10,6 +10,8 @@
 
 * `recipe()` will now show better error when columns are misspelled in formula (#1283).
 
+* `add_role()` now errors if a column would simultaneously have roles `"outcome"` and `"predictor"`. (#935)
+
 * Significant speedup in `step_dummy()` when applied to many columns. (#1305)
 
 * `recipe()` can now take data.frames with list-columns or sf data.frames as input to `data`. (#1283)